Alert

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {FlatList, View} from 'react-native';
+import {FlatList, View, Alert} from 'react-native';
 import {ListItem} from 'react-native-elements';
 import { baseUrl } from '../shared/baseurl';
 import { connect } from 'react-redux';
@@ -31,7 +31,23 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                {text: 'Calcel',
+                                onPress: () => console.log(item.name + ' Not Deleted'),
+                                style: 'cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            {cancelable: false}
+                        );
+                    } 
                 }
             ];
